- Fixed calc dinuc frequency function - linted python docs - Added multiple docstring lines

+++ gtars-python/src/models/region_set.rs
@@ -197,7 +197,7 @@ impl PyRegionSet {
         Ok(())
     }
 
-    fn region_widths(&self) -> Vec<usize> {
+    fn region_widths(&self) -> Vec<u32> {
         self.regionset.region_widths().unwrap()
     }
